#355
- Renamed cp.metadata to cp.config
- General Code Cleanup
