+++ src/cp/plugins/core/helpandsupport/userguide.lua
@@ -7,7 +7,7 @@
 --------------------------------------------------------------------------------
 -- EXTENSIONS:
 --------------------------------------------------------------------------------
-local metadata			= require("cp.metadata")
+local metadata			= require("cp.config")
 
 --------------------------------------------------------------------------------
 -- CONSTANTS:
