+++ src/cp/plugins/core/preferences/general.lua
@@ -12,7 +12,7 @@ local log				= require("hs.logger").new("preferences")
 local application		= require("hs.application")
 local console			= require("hs.console")
 
-local metadata			= require("cp.metadata")
+local metadata			= require("cp.config")
 local fcp				= require("cp.finalcutpro")
 local dialog			= require("cp.dialog")
