+++ src/cp/plugins/core/preferences/panels/plugins.lua
@@ -21,7 +21,7 @@ local toolbar                  					= require("hs.webview.toolbar")
 local webview									= require("hs.webview")
 
 local dialog									= require("cp.dialog")
-local metadata									= require("cp.metadata")
+local metadata									= require("cp.config")
 local tools										= require("cp.tools")
 
 --------------------------------------------------------------------------------
