+++ src/cp/plugins/finalcutpro/browser/playhead.lua
@@ -15,7 +15,7 @@ local timer							= require("hs.timer")
 
 local fcp							= require("cp.finalcutpro")
 local dialog						= require("cp.dialog")
-local metadata						= require("cp.metadata")
+local metadata						= require("cp.config")
 
 --------------------------------------------------------------------------------
 -- CONSTANTS:
