+++ src/cp/plugins/finalcutpro/clipboard/shared.lua
@@ -17,7 +17,7 @@ local fs										= require("hs.fs")
 local fcp										= require("cp.finalcutpro")
 local dialog									= require("cp.dialog")
 local plist										= require("cp.plist")
-local metadata									= require("cp.metadata")
+local metadata									= require("cp.config")
 local tools										= require("cp.tools")
 
 --------------------------------------------------------------------------------
