+++ src/cp/plugins/finalcutpro/export/batch.lua
@@ -10,7 +10,7 @@
 local dialog		= require("cp.dialog")
 local fcp 			= require("cp.finalcutpro")
 local just			= require("cp.just")
-local metadata		= require("cp.metadata")
+local metadata		= require("cp.config")
 local tools			= require("cp.tools")
 
 --------------------------------------------------------------------------------
