+++ src/cp/plugins/finalcutpro/hacks/playbackrendering.lua
@@ -13,7 +13,7 @@ local application		= require("hs.application")
 
 local dialog			= require("cp.dialog")
 local fcp				= require("cp.finalcutpro")
-local metadata			= require("cp.metadata")
+local metadata			= require("cp.config")
 local plist				= require("cp.plist")
 local tools				= require("cp.tools")
