+++ src/cp/plugins/finalcutpro/hud/init.lua
@@ -33,7 +33,7 @@ local ax										= require("hs._asm.axuielement")
 local plugins									= require("cp.plugins")
 local dialog									= require("cp.dialog")
 local fcp										= require("cp.finalcutpro")
-local metadata									= require("cp.metadata")
+local metadata									= require("cp.config")
 local tools										= require("cp.tools")
 local commands									= require("cp.commands")
 local template									= require("cp.template")
