+++ src/cp/plugins/finalcutpro/import/ignorecard.lua
@@ -14,7 +14,7 @@ local application				= require("hs.application")
 local timer						= require("hs.timer")
 
 local fcp						= require("cp.finalcutpro")
-local metadata					= require("cp.metadata")
+local metadata					= require("cp.config")
 
 --------------------------------------------------------------------------------
 -- CONSTANTS:
