+++ src/cp/plugins/finalcutpro/menu/timeline.lua
@@ -9,7 +9,7 @@
 --------------------------------------------------------------------------------
 -- EXTENSIONS:
 --------------------------------------------------------------------------------
-local metadata					= require("cp.metadata")
+local metadata					= require("cp.config")
 
 --------------------------------------------------------------------------------
 -- CONSTANTS:
