+++ src/cp/plugins/finalcutpro/os/touchbar.lua
@@ -15,7 +15,7 @@ local touchbar 									= require("hs._asm.touchbar")
 
 local dialog									= require("cp.dialog")
 local fcp										= require("cp.finalcutpro")
-local metadata									= require("cp.metadata")
+local metadata									= require("cp.config")
 
 --------------------------------------------------------------------------------
 -- CONSTANTS:
