+++ src/cp/plugins/finalcutpro/preferences/scanfinalcutpro.lua
@@ -14,7 +14,7 @@ local application		= require("hs.application")
 local dialog			= require("cp.dialog")
 local fcp				= require("cp.finalcutpro")
 local just				= require("cp.just")
-local metadata			= require("cp.metadata")
+local metadata			= require("cp.config")
 local tools				= require("cp.tools")
 
 --------------------------------------------------------------------------------
