+++ src/cp/welcome/init.lua
@@ -20,7 +20,7 @@ local webview									= require("hs.webview")
 
 local dialog									= require("cp.dialog")
 local fcp										= require("cp.finalcutpro")
-local metadata									= require("cp.metadata")
+local metadata									= require("cp.config")
 local template									= require("cp.template")
 local tools										= require("cp.tools")
