--- src/init.lua
@@ -1,1 +1,0 @@
-cp = require("cp")
